Move timers module into dedicated file
